Exception handling, Serilog logging, JWT auth fixes, API versioning

+++ ToDoApp.API/Controllers/TasksController.cs
@@ -1,20 +1,24 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Asp.Versioning;
+using FluentValidation;
+using FluentValidation.Results;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using ToDoApp.Core.DTO;
 using ToDoApp.Core.Entities;
 using ToDoApp.Core.Interfaces;
 using ToDoApp.Infrastructure.Data;
-using ToDoApp.Core.DTO;
 using ToDoApp.Infrastructure.Services;
-using FluentValidation;
-using FluentValidation.Results;
 
 namespace ToDoApp.API.Controllers
 {
+    [Authorize]
     [ApiController]
-    [Route("api/tasks")]
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/tasks")]
     public class TasksController : ControllerBase
     {
         private IValidator<TaskCreateDto> _createValidator;
@@ -32,16 +36,30 @@ namespace ToDoApp.API.Controllers
 
         //GET: api/tasks
         [HttpGet]
-        public async Task<IActionResult> GetTasks([FromQuery]int userId)
+        public async Task<IActionResult> GetTasks()
         {
+            //return Ok(User.Claims.Select(c => new { c.Type, c.Value }).ToList());
+
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (userIdClaim == null) return Unauthorized();
+
+            int userId = int.Parse(userIdClaim);
+
             var tasks = await _taskService.GetUserTasksAsync(userId);
             return Ok(tasks);
         }
 
         //GET: api/tasks/{id}
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetOneTask(int id, [FromQuery] int userId)
+        public async Task<IActionResult> GetOneTask(int id)
         {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            
+            if (userIdClaim == null) return Unauthorized();
+            
+            int userId = int.Parse(userIdClaim);
+
             var oneTask = await _taskService.GetTaskByIdAsync(id, userId);
 
             if (oneTask == null)
@@ -54,9 +72,16 @@ namespace ToDoApp.API.Controllers
 
         //POST: /api/tasks
         [HttpPost]
-        public async Task<ActionResult<TaskResponseDto>> CreateTask([FromQuery] int userId, [FromBody] TaskCreateDto dto)
+        public async Task<ActionResult<TaskResponseDto>> CreateTask([FromBody] TaskCreateDto dto)
         {
+
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
             
+            if (userIdClaim == null) return Unauthorized();
+            
+            int userId = int.Parse(userIdClaim);
+
+
             ValidationResult result = await _createValidator.ValidateAsync(dto);
 
             if(!result.IsValid)
@@ -71,8 +96,14 @@ namespace ToDoApp.API.Controllers
 
         //PUT: /api/tasks/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateTask(int id, [FromQuery] int userId, [FromBody] TaskUpdateDto updatedTask)
+        public async Task<IActionResult> UpdateTask(int id, [FromBody] TaskUpdateDto updatedTask)
         {
+
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (userIdClaim == null) return Unauthorized();
+
+            int userId = int.Parse(userIdClaim);
 
             ValidationResult validationResult = await _updateValidator.ValidateAsync(updatedTask);
             if(!validationResult.IsValid)
@@ -98,8 +129,14 @@ namespace ToDoApp.API.Controllers
 
         //DELETE: /api/tasks/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteTask(int id, [FromQuery] int userId)
+        public async Task<IActionResult> DeleteTask(int id)
         {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (userIdClaim == null) return Unauthorized();
+
+            int userId = int.Parse(userIdClaim);
+
             var deleted = await _taskService.DeleteTaskAsync(id, userId);
 
             if (!deleted)
